Re-define highlights on ColorScheme autocommand
When using `:colorscheme` highlights are deleted, and need to be
re-defined.

This wraps/moves this into `semshi#init()`, and triggers it on the
`ColorScheme` autocommand.

It also wraps this and the conditional definition of the other
autocommands in a new `SemshiInit` augroup, to not define the
autocommands multiple times.

Fixes https://github.com/numirias/semshi/issues/70.